requirements.txt
yfinance
requests
beautifulsoup4
google-genai